Rename and move modules, working expansion function for dia audio
